+++ sonarr/downloadclient_test.go
@@ -499,7 +499,7 @@ func TestDeleteDownloadClient(t *testing.T) {
 			t.Parallel()
 			mockServer := test.GetMockServer(t)
 			client := sonarr.New(starr.New("mockAPIkey", mockServer.URL, 0))
-			err := client.DeleteDownloadClient(test.WithRequest.(int))
+			err := client.DeleteDownloadClient(test.WithRequest.(int64))
 			assert.ErrorIs(t, err, test.WithError, "error is not the same as expected")
 		})
 	}
